reup

+++ database/migrations/2025_02_08_161356_add_user_catalogue_id_to_users.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             
-            $table->integer('user_catalogue_id')->default(2);
+            $table->bigInteger('user_catalogue_id')->default(2);
         });
     }
